Minor bug fix

+++ main.py
@@ -6,7 +6,7 @@ from nrfsim.envs.dynamic_soaring import DynamicSoaringEnv
 
 np.random.seed(1)
 
-env = DynamicSoaringEnv(initial_state=np.array([0, 0, -5, 13, 0, 0]))
+env = DynamicSoaringEnv(initial_state=np.array([0, 0, -5, 13, 0, 0]).astype('float'))
 
 time_step = 0.01
 time_series = np.arange(0, 2, time_step)
